+++ src/main/java/com/utez/integradora/repository/UserRepository.java
@@ -2,12 +2,13 @@ package com.utez.integradora.repository;
 
 import com.utez.integradora.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     Optional<UserEntity> findByEmail(String email);
     Optional<UserEntity> findByPhone(String phone);
 }
